+++ CHANGELOG.md
@@ -8,6 +8,7 @@ BREAKING CHANGES
 - [api] Switch to RPC protocol
 - [core] Fix issue with incorrect coin conversion
 - [core] Limit coins supply to 1,000,000,000,000,000
+- [core] Set minimal reserve and min/max coin supply in CreateCoin tx
 
 IMPROVEMENT